+++ planter/database/utils/duckdb_utils.py
@@ -812,7 +812,7 @@ def _create_minimal_schema() -> str:
         library_layout VARCHAR,
         instrument VARCHAR,
         run_spots INTEGER,
-        run_bases INTEGER,
+        run_bases BIGINT,
         run_published TIMESTAMP
     );
